+++ packages/article-main-standard/src/article-main-standard.web.js
@@ -84,6 +84,7 @@ class ArticlePage extends Component {
       paidContentClassName,
       isPreview,
       additionalRelatedArticlesFlag,
+      inlineRelatedArticlesFlag,
       algoliaSearchKeys,
       latestFromSectionFlag,
       latestFromSection,
@@ -108,6 +109,7 @@ class ArticlePage extends Component {
           paidContentClassName={paidContentClassName}
           isPreview={isPreview}
           additionalRelatedArticlesFlag={additionalRelatedArticlesFlag}
+          inlineRelatedArticlesFlag={inlineRelatedArticlesFlag}
           algoliaSearchKeys={algoliaSearchKeys}
           latestFromSectionFlag={latestFromSectionFlag}
           latestFromSection={latestFromSection}