+++ packages/article-skeleton/src/head.web.js
@@ -159,6 +159,16 @@ function Head({
     getThumbnailUrl(article) ||
     (getFallbackThumbnailUrl169 ? getFallbackThumbnailUrl169() : null);
 
+  const defaultAuthorSchema = {
+    "@type": "Organization",
+    name: "The Times"
+  };
+  const textByLineAuthorSchema = authorName
+    ? { "@type": "Person", name: authorName }
+    : null;
+  const authorSchema =
+    (authors && authors.length ? authors : textByLineAuthorSchema) ||
+    defaultAuthorSchema;
   const jsonLD = {
     "@context": "http://schema.org",
     "@type": "NewsArticle",
@@ -188,12 +198,9 @@ function Head({
       caption
     },
     thumbnailUrl,
-    dateModified
+    dateModified,
+    author: authorSchema
   };
-
-  if (authors && authors.length) {
-    Object.assign(jsonLD, { author: authors });
-  }
 
   const videoJsonLD = hasVideo
     ? {