+++ packages/card/__tests__/ios/card-tablet-with-style.ios.test.js
@@ -1,3 +1,0 @@
-import shared from "../tablet-with-style.native";
-
-shared();