+++ packages/gradient/__tests__/android/gradient-stroke.android.test.js
@@ -1,3 +1,0 @@
-import shared from "../shared-stroke.native";
-
-shared();