+++ packages/interactive-wrapper/__tests__/ios/interactive-wrapper.test.js
@@ -1,3 +1,0 @@
-import nativeTest from "../interactive-wrapper.native";
-
-nativeTest();