+++ packages/key-facts/__tests__/android/key-facts-with-style.android.test.js
@@ -1,3 +1,0 @@
-import shared from "../shared-key-facts-with-style.native";
-
-shared();