+++ packages/related-articles/__tests__/android/std.android-1article.test.js
@@ -1,3 +1,0 @@
-import shared from "../shared-std-1article.native";
-
-shared();