+++ packages/svgs/__tests__/ios/svgs.test.js
@@ -1,5 +1,0 @@
-import shared from "../shared";
-import viewBox from "../viewbox";
-
-shared();
-viewBox();