+++ packages/article-flag/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.10.12](https://github.com/newsuk/times-components/compare/@times-components/article-flag@2.10.11...@times-components/article-flag@2.10.12) (2021-08-09)
+
+**Note:** Version bump only for package @times-components/article-flag
+
+
+
+
+
 ## [2.10.11](https://github.com/newsuk/times-components/compare/@times-components/article-flag@2.10.10...@times-components/article-flag@2.10.11) (2021-07-05)
 
 **Note:** Version bump only for package @times-components/article-flag