+++ packages/error-view/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.60](https://github.com/newsuk/times-components/compare/@times-components/error-view@2.3.59...@times-components/error-view@2.3.60) (2021-08-09)
+
+**Note:** Version bump only for package @times-components/error-view
+
+
+
+
+
 ## [2.3.59](https://github.com/newsuk/times-components/compare/@times-components/error-view@2.3.58...@times-components/error-view@2.3.59) (2021-07-05)
 
 **Note:** Version bump only for package @times-components/error-view