+++ packages/video-label/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.64](https://github.com/newsuk/times-components/compare/@times-components/video-label@2.4.63...@times-components/video-label@2.4.64) (2021-08-09)
+
+**Note:** Version bump only for package @times-components/video-label
+
+
+
+
+
 ## [2.4.63](https://github.com/newsuk/times-components/compare/@times-components/video-label@2.4.62...@times-components/video-label@2.4.63) (2021-07-05)
 
 **Note:** Version bump only for package @times-components/video-label